+++ src/AdsenseAdvertisement.js
@@ -1,5 +1,4 @@
 var React = require("react");
-var ReactDOM = require("react-dom");
 
 var AdsenseAdvertisement = React.createClass({
   render: function() {
