+++ src/NiteHowTo.js
@@ -1,0 +1,154 @@
+var React = require("react");
+var { Thumbnail } = require("react-bootstrap");
+
+var NiteHowTo = React.createClass({
+  render: function() {
+    return (
+      <div className="howTo">
+        <p>
+          {" "}
+          本計算機ではこれまで、二手三手スキル混みの最適編成計算について、
+          総攻撃力に技巧期待値と期待攻撃回数を単に乗算した"総合*回数*技巧期待値"を用いていました。{" "}
+          <br />
+          この値を用いた場合、
+          <strong>
+            総合攻撃力が低くても攻撃回数を伸ばした方が期待される攻撃力(〜ダメージ)が大きい
+          </strong>としてソートされる傾向があります。
+        </p>
+        <p>
+          しかしながら、<a href="http://ch.nicovideo.jp/suitonjp/blomaga/ar1017708">
+            suitonさんがご提案なされた
+          </a>ように、
+          奥義を使用した場合のダメージまで考慮した1サイクルあたりのダメージをベースにソートを行った方が、
+          総合的な火力として適切ではないか、と私は考えています。
+          また、減衰補正を考慮しない場合、無制限に総合攻撃力が高い編成が優先される事になりますが、
+          実際には減衰到達後は連続攻撃で稼いでいくことになるため、総合攻撃力ベースのソートで火力を測るのは限界があると考えられます。<br />
+          そこで、本計算機においても、suitonさんご提案の1サイクルあたりダメージについて、
+          更に技巧期待値・奥義ゲージ上昇量・奥義倍率・ダメージ減衰補正を適用した上で、
+          ソートするキーとして選択できるよう追加を行いました。
+        </p>
+
+        <h3>詳しい計算手順</h3>
+        <p>まず、1サイクルあたりダメージを以下のように定義します。</p>
+        <pre>
+          1サイクルのダメージ = 奥義ダメージ +
+          (奥義ゲージが貯まるまでのターン数) * (通常攻撃ダメージ)
+        </pre>
+        <p>(本計算機ではアビリティダメージ分については考慮していません。)</p>
+        <p>ここで、DA率・TA率から、1ターンあたりに期待される攻撃回数は</p>
+        <pre>
+          期待攻撃回数 = 3 * TA率 + (1 - TA率)(2 * DA率 + 1 * (1 - DA率))
+        </pre>
+        <p>と計算できます。ここから同様に、1ターンあたりの奥義ゲージ上昇量は</p>
+        <pre>
+          期待奥義ゲージ上昇量 = TA率 * 37 + (1 - TA率)(DA率 * 22 + (1 - DA率) *
+          10)
+        </pre>
+        <p>
+          と計算できます。したがって、奥義ゲージが100%になるまでに必要なターン数は
+        </p>
+        <pre>必要ターン数 = Math.ceil(100.0 / 期待奥義ゲージ上昇量) </pre>
+        <p>となります。(Math.ceilは切り上げを意味しています。)</p>
+        <p className="text-danger">
+          2017/01/06に奥義必要ターン数を、切り捨てではなく
+        </p>
+        <pre>必要ターン数 = 100.0 / 期待奥義ゲージ上昇量 </pre>
+        <p className="text-danger">
+          上記の形へ変更しました。若干結果に変動があるかと思います。
+        </p>
+        <p>
+          ターン数については上記で求めることができましたので、次にダメージの算出を行います。
+          まず、ランク・ジョブ・武器編成などから計算された攻撃力の値(結果欄の"総合攻撃力")を用いて、
+        </p>
+        <pre> 単攻撃ダメージ(減衰補正前) = 総合攻撃力 / 敵防御固有値 </pre>
+        <p>
+          として、減衰補正前のダメージを求めることができます。
+          敵防御固有値は一般に10、硬い敵で13-15程度のようです。(防御デバフありの場合その分固有値が小さくなると考えてください。)
+        </p>
+        <p>ここで、技巧分を考慮すると、</p>
+        <pre>技巧期待値 = 技巧倍率 * 技巧スキル率 + (1.0 - 技巧スキル率) </pre>
+        <p>という式で技巧期待値が求まります。(複数ある場合にも対応)</p>
+        <p>
+          連続攻撃でクリティカルが発生した場合全てクリティカルになりますので、期待値を考える場合には単純に総合攻撃力への乗算で良いかと思われます。
+        </p>
+        <pre>
+          {" "}
+          単攻撃ダメージ(減衰補正前, 技巧補正混み) = 総合攻撃力 * 技巧期待値 /
+          敵防御固有値{" "}
+        </pre>
+        <p>次に、奥義ダメージについては</p>
+        <pre>
+          {" "}
+          奥義ダメージ(減衰補正前) = 総合攻撃力 * 奥義倍率 / 敵防御固有値{" "}
+        </pre>
+        <p>として計算できます。</p>
+        <p>
+          補正前の単攻撃ダメージが30万を超えた場合、もしくは奥義ダメージが100万を超えた場合、減衰補正を適用することになります。(参考:{" "}
+          <a href="http://greatsundome.hatenablog.com/entry/2015/10/07/114737">
+            すんどめ侍のグラブル生活 - 【グラブル】減衰補正検証まとめ
+          </a>)
+        </p>
+        <p>
+          (減衰補正については注記にも記してあるとおり、8/21現在は正確に補正されているか確認できていませんので、その点のみご注意下さい。)
+        </p>
+        <p>
+          減衰補正されたダメージ値を得られたとして、通常攻撃時の1ターンあたりに期待されるダメージは
+        </p>
+        <pre> 通常攻撃ダメージ = 期待攻撃回数 * 単攻撃ダメージ(減衰補正後)</pre>
+        <p>となります。</p>
+        <p>
+          以上で、最初に示した式中の右辺の全ての項が求まりましたので、1サイクルあたりのダメージが算出できます。
+        </p>
+        <p>これを、</p>
+        <pre>1サイクル = 奥義ゲージが貯まるまでの必要ターン数 + 1</pre>
+        <p>で割ったものが、最終的な計算結果となります。</p>
+        <pre>予想ターン毎ダメージ = 1サイクルのダメージ / 1サイクル</pre>
+        <p>
+          優先する項目を"予想ターン毎ダメージ"に変更することで、上記手順で得られた計算値を元にソートすることが可能です。
+          また、表示項目の予想ターン毎ダメージにチェックを入れることで、計算値を表示することができます。
+          (ついでに、奥義ゲージ上昇量と必要ターンについても表示できるようにしておきました。)
+        </p>
+
+        <h3>所見</h3>
+        <p>
+          これまでの"総合*回数*技巧期待値"でソートした場合の結果と、上記のソートキーを用いた場合の結果を見比べると
+          やみくもにDA・TA率が最大となるように編成されるのではなく、DA・TA率を確保した上で総合攻撃力を高める、という編成が上位にきやすくなったと感じています。
+        </p>
+        <p>
+          例えばの<a href="http://hsimyu.net/motocal/?id=748">編成例</a>ですが、下記画像のように
+        </p>
+        <Thumbnail
+          alt="総合*回数*技巧期待値の場合"
+          src="./otherImages/soukaigi.png"
+        >
+          <h2 className="text-warning">総合*回数*技巧期待値の場合</h2>
+          <p className="text-warning">
+            DA:74.1%, 奥義ゲージ上昇期待値が19.89%、必要ターンは6
+          </p>
+        </Thumbnail>
+        <Thumbnail
+          alt="予想ターン毎ダメージの場合"
+          src="./otherImages/cycleDamage.png"
+        >
+          <h2 className="text-info">予想ターン毎ダメージの場合</h2>
+          <p className="text-info">
+            DA: 67.3%, 奥義ゲージ上昇期待値が19.12%、必要ターンは6のまま
+          </p>
+        </Thumbnail>
+        <p>
+          DA率最大の編成が1位ではなくなっています。(もちろん、単なる1編成例ですので、ジョブ・DAバフ・TAバフ・奥義ゲージ上昇バフ・奥義倍率等によって、大きく結果が変わる可能性もあります。)
+        </p>
+        <p>
+          一概に後者の編成のが強いとは断言できませんが、1つの目安として使って頂ければと思います。
+        </p>
+        <p>
+          キャラ混みの平均値についても総合*回数*技巧期待値同様に計算できるようになっていますが、現在はキャラ別の奥義倍率についてはサポートしていませんので、あくまで目安としてお使い下さい。
+        </p>
+        <p>
+          また、通常攻撃ダメージの減衰補正・奥義ダメージの減衰補正についてのより詳しい情報があればご提供頂けると助かります。
+        </p>
+      </div>
+    );
+  }
+});
+module.exports.NiteHowTo = NiteHowTo;
