+++ src/RegisteredArm.js
@@ -1,365 +1,21 @@
 var React = require("react");
 var {
-  Label,
-  Nav,
-  NavItem,
-  Navbar,
-  NavDropdown,
-  MenuItem,
-  Collapse,
-  Thumbnail,
-  ControlLabel,
   Button,
-  ButtonGroup,
-  ButtonToolbar,
-  DropdownButton,
-  SplitButton,
   FormControl,
-  InputGroup,
   FormGroup,
-  Checkbox,
   Modal,
   Image,
-  Popover,
-  Col,
-  Row,
-  Grid,
   HelpBlock
 } = require("react-bootstrap");
+
 var GlobalConst = require("./global_const.js");
 var _ua = GlobalConst._ua;
 var selector = GlobalConst.selector;
 var elementTypes = GlobalConst.elementTypes;
 var skilltypes = GlobalConst.skilltypes;
+
+var { SendRequest } = require("./SendRequest.js");
 var intl = require("./translate.js");
-var { githubAPItoken } = require("./secret_consts.js");
-
-var SendRequest = React.createClass({
-  getInitialState: function() {
-    return {
-      name: "",
-      message: "",
-      yourname: "",
-      sendingMessage: false
-    };
-  },
-  sendRequestToGithub: function() {
-    if (this.state.name != "") {
-      this.setState({ sendingMessage: true });
-      var sendData = { body: "" };
-      if (this.yourname != "") {
-        sendData["body"] += "投稿者: " + this.state.yourname + "\n";
-      }
-
-      sendData["body"] +=
-        "種別: " +
-        this.props.type +
-        "\n" +
-        this.props.type +
-        "名: " +
-        this.state.name +
-        "\n";
-
-      if (this.message != "") {
-        sendData["body"] += "メッセージ:" + this.state.message + "\n";
-      }
-
-      $.ajax({
-        url:
-          "https://api.github.com/repos/hoshimi/motocal/issues/" +
-          this.props.issueNumber +
-          "/comments",
-        type: "post",
-        data: JSON.stringify(sendData),
-        beforeSend: function(xhr) {
-          xhr.setRequestHeader("Authorization", "token " + githubAPItoken);
-          xhr.setRequestHeader("Accept", "application/vnd.github.v3+json");
-        },
-        success: function(data) {
-          alert(intl.translate("送信成功", this.props.locale));
-          this.setState({ sendingMessage: false });
-          this.props.closeSendRequest();
-        }.bind(this),
-        error: function(xhr, status, err) {
-          this.setState({ sendingMessage: false });
-          this.props.closeSendRequest();
-        }.bind(this)
-      });
-    } else {
-      if (this.props.locale == "ja") {
-        alert("要望した" + this.props.type + "名が空です");
-      } else {
-        alert(this.props.type + " name must not be empty.");
-      }
-    }
-  },
-  handleEvent: function(key, e) {
-    var newState = this.state;
-    newState[key] = e.target.value;
-    this.setState(newState);
-  },
-  render: function() {
-    var locale = this.props.locale;
-
-    return (
-      <div className="sendRequestForm">
-        <p>
-          {" "}
-          <a href="https://github.com/hoshimi/motocal">
-            {" "}
-            github.com / motocal{" "}
-          </a>{" "}
-        </p>
-        <p>{intl.translate("要望送信メッセージ", locale)}</p>
-        <table className="table table-bordered">
-          <tbody>
-            <tr>
-              <th className="bg-primary">
-                {intl.translate("要望種別", locale)}
-              </th>
-              <td>{this.props.type}</td>
-            </tr>
-            <tr>
-              <th className="bg-primary">
-                {this.props.type}
-                {intl.translate("名", locale)}
-              </th>
-              <td>
-                <FormControl
-                  type="text"
-                  value={this.state.name}
-                  onChange={this.handleEvent.bind(this, "name")}
-                  style={{ textAlign: "left" }}
-                />
-              </td>
-            </tr>
-            <tr>
-              <th className="bg-primary">
-                {intl.translate("メッセージ", locale)}
-              </th>
-              <td>
-                <FormControl
-                  componentClass="textarea"
-                  value={this.state.message}
-                  onChange={this.handleEvent.bind(this, "message")}
-                  style={{ textAlign: "left", height: "200px" }}
-                />
-              </td>
-            </tr>
-            <tr>
-              <th className="bg-primary">
-                {intl.translate("あなたのお名前", locale)}
-              </th>
-              <td>
-                <FormControl
-                  type="text"
-                  value={this.state.yourname}
-                  onChange={this.handleEvent.bind(this, "yourname")}
-                  style={{ textAlign: "left" }}
-                />
-              </td>
-            </tr>
-          </tbody>
-        </table>
-        <Button
-          bsStyle="primary"
-          onClick={this.sendRequestToGithub}
-          disabled={this.sendingMessage}
-        >
-          {intl.translate("送信", locale)}
-        </Button>
-      </div>
-    );
-  }
-});
-
-var RegisteredChara = React.createClass({
-  getInitialState: function() {
-    return {
-      filterText: "",
-      filterElement: "all",
-      charaData: {},
-      limit: 50,
-      openSendRequest: false
-    };
-  },
-  componentDidMount: function() {
-    $.ajax({
-      url: "./charaData.json",
-      dataType: "json",
-      cache: false,
-      timeout: 10000,
-      success: function(data) {
-        this.setState({ charaData: data });
-      }.bind(this),
-      error: function(xhr, status, err) {
-        alert(
-          "Error!: キャラデータの取得に失敗しました。\nstatus: " +
-            status +
-            "\nerror message: " +
-            err.toString()
-        );
-        console.log("xhr:", xhr);
-        console.log("status:", status);
-        console.log("error: ", err);
-      }.bind(this)
-    });
-  },
-  clickedTemplate: function(e) {
-    this.props.onClick(this.state.charaData[e.target.getAttribute("id")]);
-  },
-  handleEvent: function(key, e) {
-    var newState = this.state;
-    newState[key] = e.target.value;
-    this.setState(newState);
-  },
-  openSendRequest: function(e) {
-    this.setState({ openSendRequest: true });
-  },
-  closeSendRequest: function(e) {
-    this.setState({ openSendRequest: false });
-  },
-  render: function() {
-    var clickedTemplate = this.clickedTemplate;
-    var filterText = this.state.filterText;
-    var filterElement = this.state.filterElement;
-    var charaData = this.state.charaData;
-    var limit = this.state.limit;
-    var displayed_count = 0;
-    var locale = this.props.locale;
-
-    if (_ua.Mobile || _ua.Tablet) {
-      return (
-        <div className="charaTemplate">
-          <span>検索:</span>
-          <FormControl
-            type="text"
-            placeholder={intl.translate("キャラ名", locale)}
-            value={this.state.filterText}
-            onChange={this.handleEvent.bind(this, "filterText")}
-          />
-          <FormControl
-            componentClass="select"
-            value={this.state.filterElement}
-            onChange={this.handleEvent.bind(this, "filterElement")}
-          >
-            {selector[locale].filterelements}
-          </FormControl>
-          <div className="charaTemplateContent">
-            {Object.keys(charaData).map(function(key, ind) {
-              var charaName = charaData[key][locale];
-              if (
-                filterElement == "all" ||
-                charaData[key].element == filterElement
-              ) {
-                if (filterText == "" || charaName.indexOf(filterText) != -1) {
-                  if (displayed_count < limit) {
-                    displayed_count++;
-                    return (
-                      <div className="onechara" key={key}>
-                        <p>{charaName}</p>
-                        <br />
-                        <Image
-                          rounded
-                          onClick={clickedTemplate}
-                          id={key}
-                          src={charaData[key].imageURL}
-                          alt={key}
-                        />
-                      </div>
-                    );
-                  } else {
-                    return "";
-                  }
-                }
-              }
-              return "";
-            })}
-          </div>
-        </div>
-      );
-    } else {
-      return (
-        <div className="charaTemplate">
-          <span>検索:</span>
-          <FormControl
-            type="text"
-            placeholder={intl.translate("キャラ名", locale)}
-            value={this.state.filterText}
-            onChange={this.handleEvent.bind(this, "filterText")}
-          />
-          <FormControl
-            componentClass="select"
-            value={this.state.filterElement}
-            onChange={this.handleEvent.bind(this, "filterElement")}
-          >
-            {selector[locale].filterelements}
-          </FormControl>
-          <div className="charaTemplateContent">
-            {Object.keys(charaData).map(function(key, ind) {
-              var charaName = charaData[key][locale];
-              if (
-                filterElement == "all" ||
-                charaData[key].element == filterElement
-              ) {
-                if (filterText == "" || charaName.indexOf(filterText) != -1) {
-                  return (
-                    <div className="onechara" key={key}>
-                      <p>{charaName}</p>
-                      <br />
-                      <Image
-                        rounded
-                        onClick={clickedTemplate}
-                        id={key}
-                        src={charaData[key].imageURL}
-                        alt={key}
-                      />
-                    </div>
-                  );
-                }
-              }
-              return "";
-            })}
-          </div>
-
-          {locale == "en" ? (
-            <p className="text-danger">
-              Help me English translation of templates on&nbsp;
-              <a
-                href="https://docs.google.com/spreadsheets/d/12R-ZQD8xy1dYYoFxjcWNN0mva1NaDIQB1XSY95jfSKg/edit"
-                target="_blank"
-              >
-                Google Spreadsheet
-              </a>!
-            </p>
-          ) : null}
-          <Button onClick={this.openSendRequest} bsStyle="danger">
-            {intl.translate("追加要望を送る", locale)}
-          </Button>
-
-          <Modal
-            show={this.state.openSendRequest}
-            onHide={this.closeSendRequest}
-          >
-            <Modal.Header closeButton>
-              <Modal.Title>
-                {intl.translate("キャラ追加要望", locale)}
-              </Modal.Title>
-            </Modal.Header>
-            <Modal.Body>
-              <SendRequest
-                locale={locale}
-                type={intl.translate("キャラ", locale)}
-                issueNumber={29}
-                closeSendRequest={this.closeSendRequest}
-              />
-            </Modal.Body>
-          </Modal>
-        </div>
-      );
-    }
-  }
-});
 
 var RegisteredArm = React.createClass({
   getInitialState: function() {
@@ -1011,5 +667,4 @@ var RegisteredArm = React.createClass({
   }
 });
 
-module.exports.RegisteredChara = RegisteredChara;
 module.exports.RegisteredArm = RegisteredArm;
