+++ src/howto.js
@@ -1,16 +1,5 @@
 var React = require("react");
-var ReactDOM = require("react-dom");
-var {
-  Thumbnail,
-  ControlLabel,
-  Button,
-  ButtonGroup,
-  FormControl,
-  Checkbox,
-  Modal,
-  Image,
-  Popover
-} = require("react-bootstrap");
+var { Thumbnail } = require("react-bootstrap");
 
 var HowTo = React.createClass({
   render: function() {
@@ -156,351 +145,4 @@ var HowTo = React.createClass({
   }
 });
 
-var HPChartHowTo = React.createClass({
-  render: function() {
-    return (
-      <Modal
-        className="hpChartTutotial"
-        show={this.props.show}
-        onHide={this.props.onHide}
-      >
-        <Modal.Header closeButton>
-          <Modal.Title>HP Chartsの使い方</Modal.Title>
-        </Modal.Header>
-        <Modal.Body>
-          <p>
-            HPチャート機能は「保存された武器編成の攻撃力等を、残りHP割合ごとに再計算する」機能です。
-          </p>
-          <h2>1.</h2>
-          <p>適当に編成を計算した後、グラフを見たい編成をグラフに加えます。</p>
-          <Thumbnail
-            alt="HPチャート操作1"
-            src="./otherImages/hpChartTutorial1.png"
-          />
-          <h2>2.</h2>
-          <p>
-            グラフに加えると、「背水渾身チャートを開く」ボタンが有効化されるので、クリックします。
-          </p>
-          <h2>3.</h2>
-          <p>
-            「優先する項目」に設定されている値を描画したグラフが表示されます。
-          </p>
-          <Thumbnail
-            alt="HPチャート操作1"
-            src="./otherImages/hpChartTutorial2.png"
-          />
-          <p className="text-danger">
-            まだサポートされていない要素が「優先する項目」に設定されている場合、"総合攻撃力"のグラフに変更されます。
-          </p>
-          <h2>4.</h2>
-          <p>上部の選択ボタンで、他の要素を表示することも可能です。</p>
-          <Thumbnail
-            alt="HPチャート操作1"
-            src="./otherImages/hpChartTutorial3.png"
-          />
-          <h2>5.</h2>
-          <p>
-            複数の召喚石組み合わせが設定されている場合、複数のグラフが作成されます。
-          </p>
-          <Thumbnail
-            alt="HPチャート操作1"
-            src="./otherImages/hpChartTutorial4.png"
-          />
-          <p className="text-danger">
-            現在は、ある組み合わせをグラフに保存すると、全てのグラフに追加されるようになっています。
-            これを召喚石別にするかどうかは、今後検討します。
-          </p>
-          <h2>注記</h2>
-          <p>
-            編成として保存されるのは「武器の組み合わせの本数」のみです。
-            そのため、武器攻撃力やバフ量などを変更した場合、結果のグラフも自動的に変更されます。
-          </p>
-          <p className="text-danger">
-            武器枠の数が追加/削除された場合、武器枠のデータがリセットされた場合は、
-            保存されている編成はリセットされてしまいますのでご注意下さい。
-            これは、武器組み合わせのみを保存しているため、誤って組み合わせで再計算されることを防ぐためです。
-          </p>
-          <p>
-            また、現在「追加した特定のグラフを削除する」機能は実装されておりませんので、
-            グラフが多くなりすぎてしまった場合、全削除を行い、保存されている編成をリセットしてください。
-          </p>
-          <p>ご要望・不具合等あればお知らせ下さい。</p>
-        </Modal.Body>
-      </Modal>
-    );
-  }
-});
-
-var NiteHowTo = React.createClass({
-  render: function() {
-    return (
-      <div className="howTo">
-        <p>
-          {" "}
-          本計算機ではこれまで、二手三手スキル混みの最適編成計算について、
-          総攻撃力に技巧期待値と期待攻撃回数を単に乗算した"総合*回数*技巧期待値"を用いていました。{" "}
-          <br />
-          この値を用いた場合、
-          <strong>
-            総合攻撃力が低くても攻撃回数を伸ばした方が期待される攻撃力(〜ダメージ)が大きい
-          </strong>としてソートされる傾向があります。
-        </p>
-        <p>
-          しかしながら、<a href="http://ch.nicovideo.jp/suitonjp/blomaga/ar1017708">
-            suitonさんがご提案なされた
-          </a>ように、
-          奥義を使用した場合のダメージまで考慮した1サイクルあたりのダメージをベースにソートを行った方が、
-          総合的な火力として適切ではないか、と私は考えています。
-          また、減衰補正を考慮しない場合、無制限に総合攻撃力が高い編成が優先される事になりますが、
-          実際には減衰到達後は連続攻撃で稼いでいくことになるため、総合攻撃力ベースのソートで火力を測るのは限界があると考えられます。<br />
-          そこで、本計算機においても、suitonさんご提案の1サイクルあたりダメージについて、
-          更に技巧期待値・奥義ゲージ上昇量・奥義倍率・ダメージ減衰補正を適用した上で、
-          ソートするキーとして選択できるよう追加を行いました。
-        </p>
-
-        <h3>詳しい計算手順</h3>
-        <p>まず、1サイクルあたりダメージを以下のように定義します。</p>
-        <pre>
-          1サイクルのダメージ = 奥義ダメージ +
-          (奥義ゲージが貯まるまでのターン数) * (通常攻撃ダメージ)
-        </pre>
-        <p>(本計算機ではアビリティダメージ分については考慮していません。)</p>
-        <p>ここで、DA率・TA率から、1ターンあたりに期待される攻撃回数は</p>
-        <pre>
-          期待攻撃回数 = 3 * TA率 + (1 - TA率)(2 * DA率 + 1 * (1 - DA率))
-        </pre>
-        <p>と計算できます。ここから同様に、1ターンあたりの奥義ゲージ上昇量は</p>
-        <pre>
-          期待奥義ゲージ上昇量 = TA率 * 37 + (1 - TA率)(DA率 * 22 + (1 - DA率) *
-          10)
-        </pre>
-        <p>
-          と計算できます。したがって、奥義ゲージが100%になるまでに必要なターン数は
-        </p>
-        <pre>必要ターン数 = Math.ceil(100.0 / 期待奥義ゲージ上昇量) </pre>
-        <p>となります。(Math.ceilは切り上げを意味しています。)</p>
-        <p className="text-danger">
-          2017/01/06に奥義必要ターン数を、切り捨てではなく
-        </p>
-        <pre>必要ターン数 = 100.0 / 期待奥義ゲージ上昇量 </pre>
-        <p className="text-danger">
-          上記の形へ変更しました。若干結果に変動があるかと思います。
-        </p>
-        <p>
-          ターン数については上記で求めることができましたので、次にダメージの算出を行います。
-          まず、ランク・ジョブ・武器編成などから計算された攻撃力の値(結果欄の"総合攻撃力")を用いて、
-        </p>
-        <pre> 単攻撃ダメージ(減衰補正前) = 総合攻撃力 / 敵防御固有値 </pre>
-        <p>
-          として、減衰補正前のダメージを求めることができます。
-          敵防御固有値は一般に10、硬い敵で13-15程度のようです。(防御デバフありの場合その分固有値が小さくなると考えてください。)
-        </p>
-        <p>ここで、技巧分を考慮すると、</p>
-        <pre>技巧期待値 = 技巧倍率 * 技巧スキル率 + (1.0 - 技巧スキル率) </pre>
-        <p>という式で技巧期待値が求まります。(複数ある場合にも対応)</p>
-        <p>
-          連続攻撃でクリティカルが発生した場合全てクリティカルになりますので、期待値を考える場合には単純に総合攻撃力への乗算で良いかと思われます。
-        </p>
-        <pre>
-          {" "}
-          単攻撃ダメージ(減衰補正前, 技巧補正混み) = 総合攻撃力 * 技巧期待値 /
-          敵防御固有値{" "}
-        </pre>
-        <p>次に、奥義ダメージについては</p>
-        <pre>
-          {" "}
-          奥義ダメージ(減衰補正前) = 総合攻撃力 * 奥義倍率 / 敵防御固有値{" "}
-        </pre>
-        <p>として計算できます。</p>
-        <p>
-          補正前の単攻撃ダメージが30万を超えた場合、もしくは奥義ダメージが100万を超えた場合、減衰補正を適用することになります。(参考:{" "}
-          <a href="http://greatsundome.hatenablog.com/entry/2015/10/07/114737">
-            すんどめ侍のグラブル生活 - 【グラブル】減衰補正検証まとめ
-          </a>)
-        </p>
-        <p>
-          (減衰補正については注記にも記してあるとおり、8/21現在は正確に補正されているか確認できていませんので、その点のみご注意下さい。)
-        </p>
-        <p>
-          減衰補正されたダメージ値を得られたとして、通常攻撃時の1ターンあたりに期待されるダメージは
-        </p>
-        <pre> 通常攻撃ダメージ = 期待攻撃回数 * 単攻撃ダメージ(減衰補正後)</pre>
-        <p>となります。</p>
-        <p>
-          以上で、最初に示した式中の右辺の全ての項が求まりましたので、1サイクルあたりのダメージが算出できます。
-        </p>
-        <p>これを、</p>
-        <pre>1サイクル = 奥義ゲージが貯まるまでの必要ターン数 + 1</pre>
-        <p>で割ったものが、最終的な計算結果となります。</p>
-        <pre>予想ターン毎ダメージ = 1サイクルのダメージ / 1サイクル</pre>
-        <p>
-          優先する項目を"予想ターン毎ダメージ"に変更することで、上記手順で得られた計算値を元にソートすることが可能です。
-          また、表示項目の予想ターン毎ダメージにチェックを入れることで、計算値を表示することができます。
-          (ついでに、奥義ゲージ上昇量と必要ターンについても表示できるようにしておきました。)
-        </p>
-
-        <h3>所見</h3>
-        <p>
-          これまでの"総合*回数*技巧期待値"でソートした場合の結果と、上記のソートキーを用いた場合の結果を見比べると
-          やみくもにDA・TA率が最大となるように編成されるのではなく、DA・TA率を確保した上で総合攻撃力を高める、という編成が上位にきやすくなったと感じています。
-        </p>
-        <p>
-          例えばの<a href="http://hsimyu.net/motocal/?id=748">編成例</a>ですが、下記画像のように
-        </p>
-        <Thumbnail
-          alt="総合*回数*技巧期待値の場合"
-          src="./otherImages/soukaigi.png"
-        >
-          <h2 className="text-warning">総合*回数*技巧期待値の場合</h2>
-          <p className="text-warning">
-            DA:74.1%, 奥義ゲージ上昇期待値が19.89%、必要ターンは6
-          </p>
-        </Thumbnail>
-        <Thumbnail
-          alt="予想ターン毎ダメージの場合"
-          src="./otherImages/cycleDamage.png"
-        >
-          <h2 className="text-info">予想ターン毎ダメージの場合</h2>
-          <p className="text-info">
-            DA: 67.3%, 奥義ゲージ上昇期待値が19.12%、必要ターンは6のまま
-          </p>
-        </Thumbnail>
-        <p>
-          DA率最大の編成が1位ではなくなっています。(もちろん、単なる1編成例ですので、ジョブ・DAバフ・TAバフ・奥義ゲージ上昇バフ・奥義倍率等によって、大きく結果が変わる可能性もあります。)
-        </p>
-        <p>
-          一概に後者の編成のが強いとは断言できませんが、1つの目安として使って頂ければと思います。
-        </p>
-        <p>
-          キャラ混みの平均値についても総合*回数*技巧期待値同様に計算できるようになっていますが、現在はキャラ別の奥義倍率についてはサポートしていませんので、あくまで目安としてお使い下さい。
-        </p>
-        <p>
-          また、通常攻撃ダメージの減衰補正・奥義ダメージの減衰補正についてのより詳しい情報があればご提供頂けると助かります。
-        </p>
-      </div>
-    );
-  }
-});
-
-var SimulatorHowTo = React.createClass({
-  render: function() {
-    return (
-      <div className="howTo">
-        <p>
-          ダメージシミュレータは<strong>
-            各ターン毎のバフ・残りHP等から予想されるダメージを計算する
-          </strong>ツールです。
-        </p>
-
-        <h2>使い方</h2>
-        <p>
-          基本的には他のグラフ機能と同様に「グラフに加えた」編成を元に計算を行います。まずは他の欄を入力し、比較したい編成を選んでおいて下さい。
-        </p>
-        <p>
-          編成をいくつかグラフに加えたら、各種グラフのボタンが有効化されます。
-          ダメージシミュレータを使う場合は「シミュレータ入力」タブを開いて下さい。
-        </p>
-        <Thumbnail
-          src="./otherImages/damage-simulator-howto-2.png"
-          href="./otherImages/damage-simulator-howto-2.png"
-        >
-          <h3>シミュレータ入力欄</h3>
-        </Thumbnail>
-
-        <p>
-          入力欄には大きく分けて「全体バフ」と「各キャラクター」のバフが存在します。
-          簡単に使いたい方は、全体バフの欄のみ弄れば各キャラクターの情報を入力しなくても問題ありません。
-          また、各キャラクターのバフは<strong>全体バフの量に加算</strong>されますのでご注意下さい。
-        </p>
-        <p>
-          （例:
-          あるターンの全体バフが通常攻刃+10%、ジータバフ欄が通常攻刃+30%となっている場合、ジータのバフは通常攻刃+40%として計算される。）
-        </p>
-        <p>
-          バフを追加したい場合、<Button>
-            <i className="fa fa-plus-square" />
-          </Button>{" "}
-          をクリックすることで欄が増えます。（現在は最大10個）
-          増やしすぎた場合は、バフ量を0%にするか、<Button>
-            <i className="fa fa-minus-square" />
-          </Button>{" "}
-          で削除して下さい。（現在は一番下にあるもののみしか削除できません。）
-          また、<Button>
-            <i className="fa fa-arrow-left" />
-          </Button>{" "}
-          や{" "}
-          <Button>
-            <i className="fa fa-arrow-right" />
-          </Button>{" "}
-          をクリックすることで、前（または次）のターンへコピーすることが可能です。
-        </p>
-        <p>
-          全てのターンのバフ情報を他のキャラ欄にコピーしたい場合は、<Button>
-            <i className="fa fa-angle-double-up" />
-          </Button>{" "}
-          または{" "}
-          <Button>
-            <i className="fa fa-angle-double-up" />
-          </Button>{" "}
-          を使用して下さい。
-        </p>
-
-        <h3>算出される値について</h3>
-        <p>
-          基本的には他のグラフと同様、技巧や二手三手の期待値から予想される攻撃力（総回技値）ですが、予想ダメージについては少し異なります。
-          最適編成算出や、他のグラフで用いている「予想ターン毎ダメージ」は、
-          "二手スキル込みの編成について"でも書いた通り、
-          <strong>
-            通常攻撃を数回行ったあと、奥義を1回打つまでの1サイクルを考えたダメージ
-          </strong>です。
-          これに対して、ダメージシミュレータで算出される「予想ダメージ」は、
-          1サイクルを考えずに、単純にそのターンに与えることが期待されるダメージです。（勿論クリティカルやDA・TA率は考慮されています）
-          これは、通常攻撃ターンと奥義ターンを設定するための変更です。
-          内部的に使っている通常攻撃ダメージ算出・奥義ダメージ算出用の関数は同一であるため、
-          単純に"1サイクルを考えるかどうか"が異なっていると考えて頂ければ結構です。
-        </p>
-
-        <h3>奥義ターンについて</h3>
-        <p>
-          欄の一番上を「通常攻撃」から「奥義」または「奥義（ダメージ無し）」に変えると、そのターンは奥義ダメージのみが与ダメージとなります。ダメージ無しの場合は0になります。
-        </p>
-        <p>
-          「奥義」と「通常攻撃」では、「奥義」の設定のが優先されます。 従って、<br />
-          全体バフが「奥義」の場合: キャラ個別の設定に関わらず、全員が奥義を打つ{" "}
-          <br />
-          全体バフが「通常」の場合:
-          「奥義」に設定されているキャラのみが奥義を打つ <br />
-          となります。
-        </p>
-        <p className="text-danger">
-          チェインバーストのダメージについてはまだサポートしておりません。そのうちやります。
-        </p>
-
-        <h3>現在未サポート</h3>
-        <p>
-          クリティカル系のバフ、追加ダメージ系のバフ、チェインバーストダメージ等
-        </p>
-
-        <h3>予想ダメージ平均の積分値について</h3>
-        <p>
-          予想ダメージを和していった値です。何ターン目でこっちの編成が……という比較のための機能です。将来的に、いくつかのバフ編成同士でも比較できたらいいな……と考えていますが、やらないかもしれません。
-        </p>
-
-        <h3>今後の予定</h3>
-        <p>
-          未サポートのバフに対応するのは勿論、なんかバフボタンを適当にドラッグ&ドロップしたら勝手に3ターン分設定されるとか、そういう感じのが欲しいですね。
-        </p>
-
-        <h3>注記</h3>
-        <p>
-          きっちりテストできていない機能ですので、不具合報告や改善点の提案などありましたらご連絡下さい。
-        </p>
-      </div>
-    );
-  }
-});
-
 module.exports.HowTo = HowTo;
-module.exports.NiteHowTo = NiteHowTo;
-module.exports.HPChartHowTo = HPChartHowTo;
-module.exports.SimulatorHowTo = SimulatorHowTo;
